refactor(mobile): created src directory and installed depdendencies
created src directory and installed depdendencies
