+++ packages/mobile/app.json
@@ -59,11 +59,6 @@
       ]
 
     ],
-    "extra": {
-      "router": {
-        "root": "./app"
-      }
-    },
     "experiments": {
       "typedRoutes": true,
       "reactCompiler": true
