--- packages/mobile/app/_layout.tsx
@@ -1,25 +1,0 @@
-import { Stack } from "expo-router";
-import { StatusBar } from "expo-status-bar";
-import "react-native-reanimated";
-
-import { GluestackUIProvider } from "@/components/ui/gluestack-ui-provider";
-import "@/global.css";
-
-export const unstable_settings = {
-	anchor: "(tabs)",
-};
-
-export default function RootLayout() {
-	return (
-		<GluestackUIProvider mode="dark">
-			<Stack>
-				<Stack.Screen name="(tabs)" options={{ headerShown: false }} />
-				<Stack.Screen
-					name="modal"
-					options={{ presentation: "modal", title: "Modal" }}
-				/>
-			</Stack>
-			<StatusBar style="auto" />
-		</GluestackUIProvider>
-	);
-}
