+++ packages/mobile/babel.config.js
@@ -11,7 +11,7 @@ module.exports = function (api) {
           root: ['./'],
 
           alias: {
-            '@': './',
+            '@': './src',
             'tailwind.config': './tailwind.config.js',
           },
         },
