+++ packages/mobile/src/app/_layout.tsx
@@ -1,6 +1,24 @@
-import "../global.css";
+import "../../global.css";
 import { Stack } from "expo-router";
+import { StatusBar } from "expo-status-bar";
+import "react-native-reanimated";
+import { GluestackUIProvider } from "@/components/ui/gluestack-ui-provider";
+
+export const unstable_settings = {
+	anchor: "(tabs)",
+};
 
 export default function RootLayout() {
-	return <Stack />;
+	return (
+		<GluestackUIProvider mode="dark">
+			<Stack>
+				<Stack.Screen name="(tabs)" options={{ headerShown: false }} />
+				<Stack.Screen
+					name="modal"
+					options={{ presentation: "modal", title: "Modal" }}
+				/>
+			</Stack>
+			<StatusBar style="auto" />
+		</GluestackUIProvider>
+	);
 }
