+++ packages/mobile/src/app/index.tsx
@@ -1,5 +1,16 @@
-import UserConsentLocationServices from "@/components/UserConsentLocationServices";
+import { Text, View } from "react-native";
+import UserConsentLocationServices from "@/components/custom/UserConsentLocationServices";
 
 export default function Index() {
-	return <UserConsentLocationServices />;
+	return (
+		<View
+			style={{
+				flex: 1,
+				justifyContent: "center",
+				alignItems: "center",
+			}}
+		>
+			<UserConsentLocationServices />
+		</View>
+	);
 }
