+++ packages/mobile/src/app/signup/index.tsx
@@ -1,6 +1,6 @@
 import { View, Text } from "react-native";
 import Auth from "../../components/custom/Auth";
-import useSession from "../hooks/useSession";
+import useSession from "../../hooks/useSession";
 
 export default function Signup() {
 	const { session } = useSession();
