+++ packages/mobile/src/assets/images/android-icon-background.png
