+++ packages/mobile/src/assets/images/android-icon-monochrome.png
