+++ packages/mobile/src/assets/images/favicon.png
