+++ packages/mobile/src/assets/images/icon.png
