+++ packages/mobile/src/assets/images/partial-react-logo.png
