+++ packages/mobile/src/assets/images/react-logo.png
