+++ packages/mobile/src/assets/images/react-logo@2x.png
