+++ packages/mobile/src/assets/images/react-logo@3x.png
