+++ packages/mobile/src/assets/images/splash-icon.png
