+++ packages/mobile/src/components/ui/gluestack-ui-provider/config.ts
