+++ packages/mobile/src/components/ui/gluestack-ui-provider/index.tsx
