+++ packages/mobile/src/components/ui/gluestack-ui-provider/script.ts
