+++ packages/mobile/src/components/ui/hstack/styles.tsx
