+++ packages/mobile/src/components/ui/text/index.tsx
