+++ packages/mobile/src/components/ui/vstack/styles.tsx
