--- packages/mobile/src/lib/utils.ts
@@ -1,10 +1,0 @@
-import { type ClassValue, clsx } from "clsx";
-import { twMerge } from "tailwind-merge";
-
-/**
- * Helper function to conditionally combine class names/styles.
- * Uses clsx for conditional grouping and twMerge for resolving Tailwind CSS conflicts.
- */
-export function cn(...inputs: ClassValue[]) {
-	return twMerge(clsx(inputs));
-}
