+++ packages/mobile/tsconfig.json
@@ -1,14 +1,14 @@
 {
   "extends": "expo/tsconfig.base",
   "compilerOptions": {
-    "baseUrl": "src",
+    "baseUrl": ".",
     "strict": true,
     // "useUnknownInCatchVariables": false,
     "isolatedModules": true, // Good to enforce monorepo safety
     "noUncheckedIndexedAccess": true, // Stricter array access
     "paths": {
       "@/*": [
-        "./*"
+        "src/*"
       ],
       "@shared/*": ["../../packages/shared/*"],
       "tailwind.config": [
